Refactor: remove unused var
Removed Api param from Handlers as it was not being used
