--- src/main/java/iudx/apd/acl/server/authentication/authorization/AuthorizationStrategy.java
@@ -1,6 +1,0 @@
-package iudx.apd.acl.server.authentication.authorization;
-
-public interface AuthorizationStrategy {
-
-  boolean isAuthorized(AuthorizationRequest authRequest);
-}
