+++ src/main/java/iudx/apd/acl/server/authentication/handler/AuthHandler.java
@@ -9,9 +9,7 @@ import io.vertx.core.Promise;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.RoutingContext;
 import iudx.apd.acl.server.authentication.AuthenticationService;
-import iudx.apd.acl.server.authentication.model.DxRole;
 import iudx.apd.acl.server.authentication.model.JwtData;
-import iudx.apd.acl.server.common.Api;
 import iudx.apd.acl.server.common.HttpStatusCode;
 import iudx.apd.acl.server.common.ResponseUrn;
 import iudx.apd.acl.server.common.RoutingContextHelper;
@@ -20,17 +18,11 @@ import org.apache.logging.log4j.Logger;
 
 public class AuthHandler implements Handler<RoutingContext> {
   static AuthenticationService authenticator;
-  static Api api;
   private final Logger LOGGER = LogManager.getLogger(AuthHandler.class);
 
-  public AuthHandler(
-      Api apis,
-      AuthenticationService authenticationService) {
+  public AuthHandler(AuthenticationService authenticationService) {
     authenticator = authenticationService;
-    api = apis;
   }
-
-
 
   @Override
   public void handle(RoutingContext context) {
