--- src/main/java/iudx/apd/acl/server/authentication/model/Method.java
@@ -1,8 +1,0 @@
-package iudx.apd.acl.server.authentication.model;
-
-public enum Method {
-  POST,
-  DELETE,
-  GET,
-  PUT;
-}
