+++ src/main/java/iudx/apd/acl/server/notification/CreateNotification.java
@@ -7,8 +7,8 @@ import static iudx.apd.acl.server.apiserver.util.Constants.STATUS_CODE;
 import static iudx.apd.acl.server.apiserver.util.Constants.TITLE;
 import static iudx.apd.acl.server.apiserver.util.Constants.TYPE;
 import static iudx.apd.acl.server.apiserver.util.Constants.USER_ID;
-import static iudx.apd.acl.server.authentication.Constants.AUD;
-import static iudx.apd.acl.server.authentication.Constants.IS_DELEGATE;
+import static iudx.apd.acl.server.authentication.util.Constants.AUD;
+import static iudx.apd.acl.server.authentication.util.Constants.IS_DELEGATE;
 import static iudx.apd.acl.server.common.HttpStatusCode.BAD_REQUEST;
 import static iudx.apd.acl.server.common.HttpStatusCode.INTERNAL_SERVER_ERROR;
 import static iudx.apd.acl.server.common.ResponseUrn.BAD_REQUEST_URN;
